requirements.txt
gradio>=3.50.2
selenium>=4.10.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-docx>=0.8.11
pandas>=2.0.3
lxml>=4.9.3